+++ src/config/file.rs
@@ -201,6 +201,8 @@ pub(super) struct SidebarFileConfig {
     pub(super) panels: Option<Vec<Vec<String>>>,
     pub(super) split: Option<bool>,
     pub(super) split_ratio: Option<f32>,
+    pub(super) background_follows_terminal: Option<bool>,
+    pub(super) gap: Option<bool>,
 }
 
 #[derive(Debug, Deserialize)]